+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - Added editor config
 
+### Changes
+ - Removed server headers
+
 ## [5.0.0] - 2021-05-23
 ### Changed
  - Updated .Net to version 5